supervisr/requirements-dev.txt
coverage
isort
pylint-django
prospector
django-debug-toolbar
sphinx
sphinx_rtd_theme
